+++ src/08-value/service.ts
@@ -6,7 +6,6 @@ import { performTransaction } from '../root/utils/performTransaction';
 import {
   ICreateValueByTypeParams,
   IDeleteValueByTypeParams,
-  IGetAllValuesByTypeParams,
   ISelectValueParams,
   IUpdateValueByTypeParams,
 } from './interfaces/services';
@@ -14,28 +13,23 @@ import { SingleValueTypes } from '../05-column/constant';
 import Column from '../models/column';
 import { IColumnDocWithType } from '../05-column/interfaces/column';
 import { checkValidType } from '../root/utils/validator';
+import validator from 'validator';
 
 export default class ValueService {
-  static async getAllValuesByType({ boardId, columnId }: IGetAllValuesByTypeParams) {
-    const foundColumn = await Column.findById(columnId).lean();
-    if (!foundColumn) throw new NotFoundError('Column is not found');
-
-    const foundAllValues = await DefaultValue.find({
-      belongBoard: { $in: [boardId, null] },
-      belongType: foundColumn.belongType,
-    })
-      .select('_id value color')
-      .lean();
-    return foundAllValues;
-  }
-
   static async createValueByType({ boardId, columnId, userId, data }: ICreateValueByTypeParams) {
+    if (!validator.isMongoId(boardId)) throw new BadRequestError(`Board Id: ${boardId} is invalid`);
+    if (!validator.isMongoId(columnId))
+      throw new BadRequestError(`Column Id: ${columnId} is invalid`);
+
+    if (!(data.hasOwnProperty('value') && data.hasOwnProperty('color')))
+      throw new BadRequestError('Missing some fields to create a new value');
+
     const foundBoard = await Board.findById(boardId);
+    if (!foundBoard) throw new NotFoundError('Board is not found');
     const foundColumn = (await Column.findById(columnId).populate({
       path: 'belongType',
       select: '_id name',
     })) as IColumnDocWithType;
-    if (!foundBoard) throw new NotFoundError('Board is not found');
     if (!foundColumn) throw new NotFoundError('Column is not found');
     if (Object.values(SingleValueTypes).includes(foundColumn.belongType.name as SingleValueTypes)) {
       throw new BadRequestError(`This type can not create more value`);
@@ -68,6 +62,9 @@ export default class ValueService {
   }
 
   static async updateValueByType({ defaultValueId, updationData }: IUpdateValueByTypeParams) {
+    if (!validator.isMongoId(defaultValueId))
+      throw new BadRequestError(`Value Id: ${defaultValueId} is invalid`);
+
     return await performTransaction(async (session) => {
       const updatedValue = await DefaultValue.findByIdAndUpdate(defaultValueId, updationData, {
         new: true,
@@ -81,15 +78,26 @@ export default class ValueService {
   }
 
   static async selectValue({ tasksColumnsId, value, valueId }: ISelectValueParams) {
+    if (!validator.isMongoId(tasksColumnsId))
+      throw new BadRequestError(`Value id of task: ${tasksColumnsId} is invalid`);
+
     const foundTasksColumns = await TasksColumns.findById(tasksColumnsId);
     if (!foundTasksColumns) throw new NotFoundError('This box is not found');
 
     if (valueId) {
+      if (foundTasksColumns.typeOfValue === 'single')
+        throw new BadRequestError('Invalid data transmitted');
+      if (!validator.isMongoId(valueId))
+        throw new BadRequestError(`Value Id: ${valueId} is invalid`);
+
       const foundDefaultValue = await DefaultValue.findById(valueId).lean();
       if (!foundDefaultValue) throw new NotFoundError(`Default value of ${valueId} is not found`);
       foundTasksColumns.valueId = foundDefaultValue._id;
       return await foundTasksColumns.save();
     }
+
+    if (foundTasksColumns.typeOfValue === 'multiple')
+      throw new BadRequestError('Invalid data transmitted');
 
     const foundColumnWithType = (await Column.findById(foundTasksColumns.belongColumn)
       .populate({
@@ -111,6 +119,11 @@ export default class ValueService {
   }
 
   static async deleteValueByType({ columnId, defaultValueId }: IDeleteValueByTypeParams) {
+    if (!validator.isMongoId(defaultValueId))
+      throw new BadRequestError(`Value Id: ${defaultValueId} is invalid`);
+
+    if (!validator.isMongoId(columnId))
+      throw new BadRequestError(`Column Id: ${columnId} is invalid`);
     return await performTransaction(async (session) => {
       const deletedValue = await DefaultValue.findByIdAndDelete(defaultValueId, { session });
       if (!deletedValue) throw new NotFoundError('Value is not found');
